@@ -2,7 +2,6 @@ package chezmoi
 
 import (
 	"bytes"
-	"context"
 	"io"
 	"io/fs"
 	"testing"
@@ -35,7 +34,7 @@ func TestZIPWriterSystem(t *testing.T) {
 			"symlink_symlink": ".dir/subdir/file\n",
 		},
 	}, func(fileSystem vfs.FS) {
-		ctx := context.Background()
+		ctx := t.Context()
 		system := NewRealSystem(fileSystem)
 		s := NewSourceState(
 			WithBaseSystem(system),
